requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
pipreqs==0.4.11
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
yarg==0.1.9
